+++ clang/lib/Driver/ToolChains/Clang.cpp
@@ -6225,6 +6225,13 @@ void Clang::ConstructJob(Compilation &C, const JobAction &JA,
   if (Arg *A = Args.getLastArg(options::OPT_finsert_intraobject_tripwires_EQ))
     CmdArgs.push_back(Args.MakeArgString(
         Twine("-finsert-intraobject-tripwires=") + A->getValue()));
+  if (Arg *A = Args.getLastArg(options::OPT_foptimize_intraobject_tripwires_EQ))
+    CmdArgs.push_back(Args.MakeArgString(
+        Twine("-foptimize-intraobject-tripwires=") + A->getValue()));
+
+  Args.AddLastArg(CmdArgs,
+                  options::OPT_mpreiniticv,
+                  options::OPT_mno_preiniticv);
 
   const XRayArgs &XRay = TC.getXRayArgs();
   XRay.addArgs(TC, Args, CmdArgs, InputType);