+++ clang/test/Layout/intraobject-tripwires-all.c
@@ -1,0 +1,2 @@
+// RUN: %clang -std=c2x -finsert-intraobject-tripwires=all -DTRIPWIRES_ALL %s
+#include "intraobject-tripwires.inc"