+++ clang/test/Layout/intraobject-tripwires-attr.c
@@ -1,0 +1,2 @@
+// RUN: %clang -std=c2x -finsert-intraobject-tripwires=attr -DTRIPWIRES_ATTR %s
+#include "intraobject-tripwires.inc"