+++ clang/test/Layout/intraobject-tripwires-none.c
@@ -1,0 +1,2 @@
+// RUN: %clang -std=c2x -finsert-intraobject-tripwires=none -DTRIPWIRES_NONE %s
+#include "intraobject-tripwires.inc"