+++ llvm/include/llvm/Bitcode/LLVMBitCodes.h
@@ -712,6 +712,8 @@ enum AttributeKindCodes {
   ATTR_KIND_FNRETTHUNK_EXTERN = 84,
   ATTR_KIND_SKIP_PROFILE = 85,
   ATTR_KIND_MEMORY = 86,
+  ATTR_KIND_CCPREINITICV = 87,
+  ATTR_KIND_CCTRIPWIRES = 88,
 };
 
 enum ComdatSelectionKindCodes {