+++ llvm/lib/Bitcode/Reader/BitcodeReader.cpp
@@ -2066,6 +2066,10 @@ static Attribute::AttrKind getAttrFromCode(uint64_t Code) {
     return Attribute::Hot;
   case bitc::ATTR_KIND_PRESPLIT_COROUTINE:
     return Attribute::PresplitCoroutine;
+  case bitc::ATTR_KIND_CCPREINITICV:
+    return Attribute::CCPreInitICV;
+  case bitc::ATTR_KIND_CCTRIPWIRES:
+    return Attribute::CCTripwires;
   }
 }