+++ llvm/lib/TargetParser/X86TargetParser.cpp
@@ -533,6 +533,7 @@ constexpr FeatureBitset ImpliedFeaturesWBNOINVD = {};
 constexpr FeatureBitset ImpliedFeaturesVZEROUPPER = {};
 constexpr FeatureBitset ImpliedFeaturesX87 = {};
 constexpr FeatureBitset ImpliedFeaturesXSAVE = {};
+constexpr FeatureBitset ImpliedFeaturesPREINITICV = {};
 
 // Not really CPU features, but need to be in the table because clang uses
 // target features to communicate them to the backend.